Moved Contentful ContentTypes into its own enum folder.
